Avoid mutating in CFG for ivar T.let autocrorect
We can avoid mutating the tree in CFG for this case by pulling the logic
forward into resolver, where the tree is still mutable.

I'm working towards making the tree immutable in `typecheckOne`, so that
we can cache the resolved tree for IDE queries, and this is one of the
few places where builder_walk.cc is mutating the input tree.